Use hazelcast-all JARs in server compatibility tests (#56)
We have to include hazelcast-sql in classpath to run SQL related
tests. Easiest way to do it is to use hazelcast-all JAR instead
of hazelcast JAR. This is what we do for the clients and now
the same logic is applied here.

Note that, it won't be a problem in 4.2- server versions as well.
We can use the hazelcast-all there too.