Revert Node.js version to 14
This is a required change, as in Node.js 10.4, the tests were failing
due to client not being able to close connections.

The client was relying on the fact that writes to an already closed
socket would throw an error and the client would be able to close the
that connection, but that is not happening on Node.js 10.4.

Looking at the recent changes done by Serkan on the Node.js client,
it seems it is possible to work around that problem but that requires
a change in the client code, and we cannot do that for already
released clients.

So, we have to revert the Node.js version back and look for long-term
solutions once Serkan is back.