Generate tests-jar for 5.2.0 (#42)
We don't generate the tests JAR anymore in 5.2.0. But, we haven't
updated the client compatibility tests to drop the dependency on
it.

This PR adds the git diff to be patch the enterprise code so that
it also generates the tests JAR.

We should revert these changes once we implement the proper solution.