Use generated certs.jar instead of enterprise-tests.jar [API-1766] (#46)
* Use generated certs.jar instead of enterprise-tests.jar

We no longer generate the enterprise-tests.jar, starting from 5.2.0.

As a solution, we have decided to generate the certificates used
in tests in a separate repo, and use the certificates from there.

We can easily do that for the new tests, but for the tests here,
we have to continue pretending as if there is an enterprise-tests
jar because we cannot change the run scripts of the already released
clients.

Luckily, all of the scripts are smart enough to not try to download
a jar, if it exists locally. We can make use of this and pretend
like the certs.jar is the enterprise-tests.jar. We have produced
certs.jar to be a drop-in replacement for the enterprise-tests.jar
and it has the certificates used by clients in the same path
as the enterprise-tests.jar.

In this PR, we download the cert.jar from the private-test-artifacts
repo, rename it as necessary, and start the remote controller
with it, hoping that the run scripts of the clients will no longer
try to download the existing jar.

* run the renaming command on the certs directory