Fix C# tests post EE/OSS patch changes (#112)
The C# client has a build script - `hz.ps1` - which is used to
build/test the codebase, and is part of the repository.

Because `client-compatibility-tests` tests against old, tagged versions
of the `hazelcast-csharp-client` repository, the script contained in
that version of the repository may not be suitable - so
`client-compatibility-tests` has it's own, updated copies of those
scripts to use for older clients.

The script in `hazelcast-csharp-client` was _partially_ updated to
support the OS/EE patch changes:
- https://github.com/hazelcast/hazelcast-csharp-client/pull/903
- https://github.com/hazelcast/hazelcast-csharp-client/pull/905

However, `client-compatibility-tests`'s versions were not updated to
suit. The obvious solution is to remove this duplication to avoid issues
like this in future.

Changes:
- Removed local copies of the script:
- `utils/net/hz-latest.ps1` - this was functionally equivalent to
`hazelcast-csharp-client`'s `hz.ps1`, except for the new changes that
hadn't been backported
- `utils/net/no-java-check/hz.ps1` - this was _related_ to an older
version of the script, but with a Java 15 check removed - this is
redundant [since `hazelcast-remote-controller` was updated to support
Java
17](https://github.com/hazelcast/hazelcast-remote-controller/pull/72)
- `utils/net/utils.ps1` - this was functionally equivalent to the latest
script in `hazelcast-csharp-client`
- now uses the latest scripts, sourced directly from
`hazelcast-csharp-client`'s `master`, for all executions (i.e. any tag -
new or old)
- add credentials required to access the OSS `SNAPSHOT` repository

In addition, further issues were found with the script, which due to the
now-centralised nature can be fixed upstream:
- https://github.com/hazelcast/hazelcast-csharp-client/pull/908
- https://github.com/hazelcast/hazelcast-csharp-client/pull/909

To test this, I ran against [my branch of `hazelcast-csharp-client` with
all of these
fixes](https://github.com/JackPGreen/hazelcast-csharp-client/tree/all-the-fixes)
(as they are not yet merged into `master`).

[In those
builds](https://github.com/hazelcast/client-compatibility-suites/actions/runs/9910419327),
all of the _builds_ now pass.
However, some of the tests fail - after re-running some additional tests
passed and others failed - which makes me think that there are are
additional flaky tests outside the scope of the original issue.