Fix C# tests post `set-output` changes
Since https://github.com/hazelcast/client-compatibility-suites/pull/178, the C# tests have been failing to start.

This is because the updated command is for `bash`, but the executing shell for the `setup_csharp_client_matrix` job is implicitly `pwsh`.

We don't actually _need_ to use Windows for the setup step, so rather than rewriting the command it's better to use Ubuntu which is less GitHub Actions resource consuming.

We _already_ do this in `csharp_client_compatibility`