Fix C# tests post `set-output` changes [DI-565] (#182)
Since https://github.com/hazelcast/client-compatibility-suites/pull/178,
the C# tests have been failing to start.

This is because the updated command is for `bash`, but the executing
shell for the `setup_csharp_client_matrix` job is implicitly `pwsh`.

We don't actually _need_ to use Windows for the setup step, so rather
than rewriting the command it's better to use Ubuntu which is less
GitHub Actions resource consuming.

We _already_ do this in `csharp_client_compatibility`.

[Example
execution](https://github.com/hazelcast/client-compatibility-suites/actions/runs/16140232371).

Fixes: [DI-565](https://hazelcast.atlassian.net/browse/DI-565)

[DI-565]:
https://hazelcast.atlassian.net/browse/DI-565?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ